+++ src/components/register/ActivationConfirmation.tsx
@@ -5,6 +5,7 @@ import {RegisterRequestClient, RegisterRequestEmployee} from "@/types/auth.ts";
 import {registerClient, registerEmployee} from "@/api/auth.ts";
 import { Button } from "@/components/ui/button";
 import {CardFooter} from "@/components/ui/card.tsx";
+import { useNavigate } from "react-router-dom";
 
 interface ActivationConfirmationProps extends React.ComponentProps<"div"> {
     onContinue?: () => void;
@@ -13,6 +14,15 @@ interface ActivationConfirmationProps extends React.ComponentProps<"div"> {
 }
 
 export default function ActivationConfirmation({ onContinue, registerdata, className, ...props}: ActivationConfirmationProps) {
+
+    const navigate = useNavigate();
+
+    const handleClick = () => {
+
+        navigate("/create-account", { state: { step2: 2 } });
+
+    };
+
     // Type guard function
     function isEmployee(data: RegisterRequestEmployee | RegisterRequestClient): data is RegisterRequestEmployee {
         return (data as RegisterRequestEmployee).department !== undefined;
@@ -41,6 +51,10 @@ export default function ActivationConfirmation({ onContinue, registerdata, class
 
     }
 
+    const handleContinue = () => {
+        navigate("/create-account");
+    };
+
     return (
         <div>
             <EmailConfirmation title="Activation email has been sent!"
@@ -57,7 +71,8 @@ export default function ActivationConfirmation({ onContinue, registerdata, class
                                 {...props}
                                 footer={
                                     <CardFooter className="flex flex-row gap-4 justify-center">
-                                        <Button variant="default" onClick={onContinue}>Continue bank account creation</Button>
+                                        <Button variant="default" onClick={handleClick}>Continue bank account creation</Button>
+
                                     </CardFooter>
                                  }>
             </EmailConfirmation>
